backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-aiplatform==1.38.1
vertexai==0.0.1
gunicorn==21.2.0
flask-limiter==3.5.0
flask-talisman==1.1.0 